US30: List living married
