+++ constants.py
@@ -1,0 +1,21 @@
+# columns for Families
+FAMILY_COLUMNS = ["ID", "Married", "Divorced", "Husband ID", "Husband Name", "Wife ID", "Wife Name", "Children"]
+
+# columns for Individuals
+INDIVIDUAL_COLUMNS = ["ID", "Name", "Gender", "Birthday", "Age", "Alive", "Death", "Child", "Spouse"]
+
+# Maps the month date with a number for date comparisons
+month_dict = {
+        'JAN' : 1,
+        'FEB' : 2,
+        'MAR' : 3,
+        'APR' : 4,
+        'MAY' : 5,
+        'JUN' : 6,
+        'JUL' : 7,
+        'AUG' : 8,
+        'SEP' : 9,
+        'OCT' : 10,
+        'NOV' : 11,
+        'DEC' : 12
+    }
